VFS and Tenant Resource Provision on AWS (#229)
* WIP

* Finished auth fetcher

* Replace URL with URI

* Finished the presigning URL

* Added support for upload URLS

* Added iaas-role to cognito groups

* Replaced Collectors.toList with toList

* Removed empty spaces

* Removed empty space

* fixed compilation changes

* Removed redundant aws factory interface

* removed tech-debt

* Compilation issues

* Decouples deployment region from service specific regions

* Refactored the Tenant service

* Refactored the Tenant management to be decoupled and open room for more features in future

* Created a put method in Tenant register

* Made all AWS services idempotent

* Compilation errors fixed

* Consistent naming

* BugFixes

* consistent naming

* Added some unit tests

* - Add IaasRolePolicyAttachmentServiceTests
- Refactored != null filters to use Objects::nonNull

* Refactored delete count to use long instead of int

* Tests for IaasPolicyService

* Formatting fixes

* Added unit tests

* Added property for eclipse incremental annotation processor

* Added tests for AwsIamPolicyClient

* Refactored the User Service

* Fixed the service auto configuration. no compilation errors or test failures except context loads

* Wired up all the beans and validated the changes

* Added decorator to the crud controller for the User Dto

* Fix GetSkus

* Added more unit tests.

Refactored the GET and DELETE APIs to not require Content-Type header

* Added unit tests

* Type fix

* Increased code coverage

* Increased code coverage

* Added unit tests for the service AWS classes

* Added Conditional bean annotation

* Added unit tests for mappers and models

* Added tests for models

* Fixed tests

* Moved the AWS specific beans to AwsConfiguration

* Refactored the Mapping annotations to be consistent across the entire application

* Refactored the code to remove extra newline

* removed unused code

* Removed the TenantService dependency and put the correct conditional

+++ src/main/java/io/company/brewcraft/service/CrudService.java
@@ -7,14 +7,13 @@ import java.util.Set;
 import io.company.brewcraft.model.Identified;
 
 public interface CrudService<ID, E extends Identified<ID>, BE, UE extends Identified<ID>, A> {
-
     boolean exists(Set<ID> ids);
 
     boolean exist(ID id);
 
-    int delete(Set<ID> ids);
+    long delete(Set<ID> ids);
 
-    int delete(ID id);
+    long delete(ID id);
 
     E get(ID id);
