+++ src/main/java/io/company/brewcraft/service/MixtureMaterialPortionService.java
@@ -11,7 +11,6 @@ import io.company.brewcraft.model.MixtureMaterialPortion;
 import io.company.brewcraft.model.UpdateMixtureMaterialPortion;
 
 public interface MixtureMaterialPortionService {
-
     public Page<MixtureMaterialPortion> getMaterialPortions(Set<Long> ids, Set<Long> mixtureIds, Set<Long> materialLotIds, Set<Long> brewStageIds, Set<Long> brewIds, int page, int size, SortedSet<String> sort, boolean orderAscending);
 
     public MixtureMaterialPortion getMaterialPortion(Long materialPortionId);
@@ -24,7 +23,7 @@ public interface MixtureMaterialPortionService {
 
     public List<MixtureMaterialPortion> patchMaterialPortions(List<UpdateMixtureMaterialPortion> materialPortion);
 
-    public int deleteMaterialPortions(Set<Long> materialPortionIds);
+    public long deleteMaterialPortions(Set<Long> materialPortionIds);
 
     public boolean materialPortionExists(Long materialPortionId);
