+++ src/main/java/io/company/brewcraft/service/MixtureRecordingService.java
@@ -11,7 +11,6 @@ import io.company.brewcraft.model.MixtureRecording;
 import io.company.brewcraft.model.UpdateMixtureRecording;
 
 public interface MixtureRecordingService {
-
     public Page<MixtureRecording> getMixtureRecordings(Set<Long> ids, Set<Long> mixtureIds, Set<Long> brewStageIds, Set<Long> brewIds, int page, int size, SortedSet<String> sort, boolean orderAscending);
 
     public MixtureRecording getMixtureRecording(Long mixtureRecordingId);
@@ -22,7 +21,7 @@ public interface MixtureRecordingService {
 
     public List<MixtureRecording> patchMixtureRecordings(List<UpdateMixtureRecording> mixtureRecording);
 
-    public int deleteMixtureRecordings(Set<Long> mixtureRecordingIds);
+    public long deleteMixtureRecordings(Set<Long> mixtureRecordingIds);
 
     public boolean mixtureRecordingExists(Long mixtureRecordingId);
