+++ src/main/java/io/company/brewcraft/service/impl/ShipmentService.java
@@ -144,12 +144,12 @@ public class ShipmentService extends BaseService implements CrudService<Long, Sh
     }
 
     @Override
-    public int delete(Set<Long> ids) {
+    public long delete(Set<Long> ids) {
         return this.repoService.delete(ids);
     }
 
     @Override
-    public int delete(Long id) {
+    public long delete(Long id) {
         return this.repoService.delete(id);
     }
