+++ src/test/java/io/company/brewcraft/service/impl/InvoiceServiceTest.java
@@ -158,10 +158,10 @@ public class InvoiceServiceTest {
 
    @Test
    public void testDelete_CallsRepoServiceDeleteBulk_WhenInvoiceExists() {
-       doReturn(123).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
+       doReturn(123L).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
 
-       final int count = this.service.delete(Set.of(1L, 2L, 3L));
-       assertEquals(123, count);
+       final long count = this.service.delete(Set.of(1L, 2L, 3L));
+       assertEquals(123L, count);
    }
 
    @Test
