+++ src/test/java/io/company/brewcraft/service/impl/PurchaseOrderServiceTest.java
@@ -137,10 +137,10 @@ public class PurchaseOrderServiceTest {
 
     @Test
     public void testDelete_CallsRepoServiceDeleteBulk_WhenPurchaseOrderExists() {
-        doReturn(123).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
+        doReturn(123L).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
 
-        final int count = this.service.delete(Set.of(1L, 2L, 3L));
-        assertEquals(123, count);
+        final long count = this.service.delete(Set.of(1L, 2L, 3L));
+        assertEquals(123L, count);
     }
 
     @Test
